requirements.txt
streamlit 
transformers=BertModel,BertTokenizer
Pytorch
base64